requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiokafka==0.8.1
aiosignal==1.3.1
alembic==1.8.1
altair==5.1.1
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.3.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
autopage==0.5.1
azure-common==1.1.28
azure-core==1.29.3
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
click-params==0.3.0
cliff==4.3.0
cloudpickle==2.2.1
cmaes==0.10.0
cmd2==2.4.3
colorlog==6.7.0
comm==0.1.4
commonmark==0.9.1
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
Cython==0.29.31
databricks-cli==0.17.7
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
docker==6.1.3
ecdsa==0.18.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.89.1
fastapi-utils==0.2.1
fastjsonschema==2.18.0
Flask==2.3.3
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
gevent==23.9.0.post1
geventhttpclient==2.0.2
gitdb==4.0.10
GitPython==3.1.34
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.58.0
gunicorn==20.1.0
h11==0.14.0
httplib2==0.19.1
httptools==0.6.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==7.8.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.5
jupyterlab_server==2.24.0
kafka-python==2.0.2
kiwisolver==1.4.5
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mistune==3.0.1
mlflow==2.4.0
mlserver==1.3.5
mlserver-mlflow==1.3.4
monotonic==1.6
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.25.2
oauthlib==3.2.2
optuna==2.10.0
orjson==3.8.14
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
plotly==5.17.0
prettytable==3.9.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-grpc-prometheus==0.7.0
pyarrow==9.0.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
PyMySQL==1.0.3
pyparsing==2.4.7
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-rapidjson==1.11
python-terraform==0.10.1
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
querystring-parser==1.2.4
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rpds-py==0.10.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
semver==3.0.1
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.22.0
starlette-exporter==0.16.0
stevedore==5.1.0
streamlit==1.27.0
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
tritonclient==2.36.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.17.6
uvloop==0.17.0
validators==0.18.2
watchgod==0.8.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==2.3.7
widgetsnbextension==3.6.5
xgboost==1.5.2
yarl==1.9.2
zenml==0.44.1
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
